refactor: move IpReputation to StrEnum in enums.py
